requirements.txt
requests
schedule
pygame
PyYAML